fixed observedOverExpected for the new version of numpy

+++ mirnylib/numutils_new.pyx
@@ -89,7 +89,7 @@ def logbinsnew(a, b, ratio=0, N=0):
     if ratio != 0:
         if N != 0:
             raise ValueError("Please specify N or ratio")
-        N = np.log(b / a) / np.log(ratio)
+        N = int(np.log(b / a) / np.log(ratio))
     elif N == 0:
         raise ValueError("Please specify N or ratio")
     data10 = np.logspace(a10, b10, N)
